Added command shortcuts and help to GitStatus
I added several more shortcuts to the GitStatus buffer

 - r = call GitRm on the file.

 - c = change the window into a commit window

 - q = close the window

 - ? = toggle showing the help

When GitStatus is invoked, there is a small message saying "type ? for
help".  Typing ? shows a comprehensive list of the shortcuts.  The help
can be toggled on and off using the ? key.

Since GitStatus's help documentation is now directly in the window, I
removed the under-documented "In git-status buffer" section at the end
of the README.doc

It's also slighly easier to run the commands now.  The cursor no longer
needs to be directly on top of a filename.  It can be anywhere on a
line containing a file.