Fixed GitCommit and added a passing test
GitCommit wasn't working properly.  I added a test and then played with
the code until it worked.  I had to comment out the support for handing
GitCommit a message directly.  It feels like maybe there was a typo in
the code anyhow, so I added a message in the code mentioning that.