Added a test for using 'r' to remove in GitStatus
The test uses a shell command to remove the file in the filesystem, then
uses the 'r' key in GitStatus to remove the file from revision control.

This type of thing happens sometimes when you use the other tools to
remove files (like if you remove them using NERDTree or a shell script)
but they are still being tracked by git.