Refactored test setup and added GitVimDiff test
Adding the GitVimDiff test made it clear that I needed to refactor the
setup files a little bit to keep them DRY but allow specific
functionality for each different category of test.

I also added a test for the recently merged GitVimDiff function.

Finally, I added a README file about testing to hopefully jumpstart
the involvement of other developers in the testing process.