Fixed GitAdd option parsing regression
At one point I was stripping the options of their prefix, then realized
that was a mistake.  I reverted to *not* stripping them, but forgot to
revert one line in the GitAdd function.

Unfortunately I'm not exactly sure how to add a test for this regression
so it doesn't happen again, since GitAdd -i executes git add as a
shell command (which I //believe// makes it impossible to automate
without something like "expect").  Anyhow, I'm going to leave it
untested for the time being.