Add a diff window for commits from GitStatus
I like to see a diff when I commit, so if g:git_always_verbose_commit=1
is set, open a window with the output of GitDiff --staged and then
open the commit message on top of that.

Note that passing --verbose to git ci doesn't work. The extra text and
diff are included in the commit message. (On Ubuntu 10.10, git 1.7.1.)